requirements.txt
Django==4.2.0
djangorestframework==3.15.0
django-cors-headers==4.0.0
psycopg2-binary==2.9.6
PyJWT==2.7.0
drf-yasg==1.21.4
python-dotenv==1.0.0
djangorestframework-simplejwt==5.5.1
drf-nested-routers==0.95.0
flake8